+++ open-media-match/src/OpenMediaMatch/templates/index.html.j2
@@ -29,7 +29,7 @@
             </tr>
         {% endfor %}
         <h1>File Upload </h1>
-        <form action = "/" method = "post" enctype="multipart/form-data">  
+        <form action = "/ui/query" method = "post" enctype="multipart/form-data">  
             <input type="file" name="file" />  Content Type (*required) 
             <select id="media" name="media">
                 {% for c in content %}
